Exception split to own files - Old namespaces aliased
