+++ src/Kdyby/RabbitMq/Exception/Exception.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Kdyby\RabbitMq\Exception;
+
+
+/**
+ * Common interface for caching github exceptions
+ *
+ * @author Filip Procházka <filip@prochazka.com>
+ */
+interface Exception
+{
+
+}
