+++ src/Kdyby/RabbitMq/Exception/InvalidArgumentException.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types = 1);
+
+namespace Kdyby\RabbitMq\Exception;
+
+
+/**
+ * @author Filip Procházka <filip@prochazka.com>
+ */
+class InvalidArgumentException extends \InvalidArgumentException implements \Kdyby\RabbitMq\Exception\Exception
+{
+
+}
