+++ src/Kdyby/RabbitMq/Exception/QueueNotFoundException.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types = 1);
+
+namespace Kdyby\RabbitMq\Exception;
+
+
+/**
+ * @author Alvaro Videla <videlalvaro@gmail.com>
+ */
+class QueueNotFoundException extends \RuntimeException implements \Kdyby\RabbitMq\Exception\Exception
+{
+
+}
